Refactor around title functions and gathering pages for sections
